# docs/swagger.yaml
basePath: /api/v1
definitions:
  Controllers.UserLogin:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  Models.Note:
    properties:
      content:
        type: string
      id:
        type: integer
      isDelete:
        type: boolean
      title:
        type: string
    type: object
  Models.Response:
    properties:
      code:
        description: Response Code
        type: integer
      data:
        description: Response Data
      status:
        description: Response Status
        type: string
    type: object
  Models.User:
    properties:
      id:
        type: integer
      isDelete:
        type: boolean
      password:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
info:
  contact:
    email: support@somewhere.com
    name: API Support
    url: http://somewhere.com/support
  description: Test Golog Swagger
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://somewhere.com/
  title: Teat Golang Swagger API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login
      operationId: Login
      parameters:
      - description: Username & Password
        in: body
        name: LoginForm
        required: true
        schema:
          $ref: '#/definitions/Controllers.UserLogin'
      produces:
      - application/json
      responses: {}
      summary: Login
      tags:
      - Authentication
  /note:
    get:
      consumes:
      - application/json
      description: Get all notes
      operationId: GetAllNotes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Models.Response'
      summary: GET All Notes
      tags:
      - Note
    post:
      consumes:
      - application/json
      description: Create Note
      operationId: CreateNote
      parameters:
      - description: Note data to be create
        in: body
        name: Note
        required: true
        schema:
          $ref: '#/definitions/Models.Note'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Models.Response'
      summary: Create Note
      tags:
      - Note
  /note/{id}:
    delete:
      consumes:
      - application/json
      description: Delete Note
      operationId: DeleteNote
      parameters:
      - description: id of note to be delete
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Delete Note
      tags:
      - Note
    get:
      consumes:
      - application/json
      description: Get NoteBy Id
      operationId: GetNoteById
      parameters:
      - description: id of note
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Models.Response'
      summary: Get NoteBy Id
      tags:
      - Note
    put:
      consumes:
      - application/json
      description: Update Note
      operationId: UpdateNote
      parameters:
      - description: id of note to be update
        in: path
        name: id
        required: true
        type: string
      - description: Note data to update
        in: body
        name: Note
        required: true
        schema:
          $ref: '#/definitions/Models.Note'
      produces:
      - application/json
      responses: {}
      summary: Update Note
      tags:
      - Note
  /user:
    get:
      consumes:
      - application/json
      description: Get All Users
      operationId: GetAllUsers
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Get All Users
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Create User
      operationId: CreateUser
      parameters:
      - description: User data to create
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/Models.User'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Create User
      tags:
      - User
  /user/{id}:
    delete:
      consumes:
      - application/json
      description: Delete User
      operationId: DeleteUser
      parameters:
      - description: id of user to be delete
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Delete User
      tags:
      - User
    get:
      consumes:
      - application/json
      description: Get UserBy Id
      operationId: GetUserById
      parameters:
      - description: id of user
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Get UserBy Id
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update User
      operationId: UpdateUser
      parameters:
      - description: id of user to be update
        in: path
        name: id
        required: true
        type: string
      - description: User data to update
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/Models.User'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Update User
      tags:
      - User
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"